requirements.txt
discord.py
box
toml
imageio[py]
numpy
pyopencl
fuzzywuzzy
python-dotenv
colorzero
python-Levenshtein